Refactor: Extract new mod `binstall`
